+++ library/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/create-effect-aware-form';
+export * from './lib/form.interface';
 export * from './lib/form-effect.type';
 export * from './lib/potentially-stable-form.test-helper';