+++ library/src/lib/create-effect-aware-form.ts
@@ -1,11 +1,10 @@
 import { merge, Observable } from 'rxjs';
-import { distinctUntilChanged, shareReplay } from 'rxjs/operators';
+import { distinctUntilChanged, map, shareReplay, startWith } from 'rxjs/operators';
 
-import { FormGroup } from '@angular/forms';
-import { map, startWith } from 'rxjs/operators';
+import { Form } from './form.interface';
 import { FormEffect } from './form-effect.type';
 
-export const createEffectAwareForm = <T extends FormGroup>(
+export const createEffectAwareForm = <T extends Form>(
   control: T,
   effects: FormEffect<T>[]
 ): Observable<T> => {